feat: 카카오 소셜 로그인 구현 및 최초 로그인 시 자동 회원가입 로직 추가

+++ src/main/java/com/ours/community/controller/MockPostController.java
@@ -35,14 +35,11 @@ public class MockPostController {
     }
 
     @GetMapping("users/{userId}")
-    public ResponseEntity<List<Post>> getPostsByUser(@PathVariable Long userId) {
+    public ResponseEntity<List<Post>> getPostsByUser(@PathVariable Integer userId) {
         User user = userRepository.findById(userId)
                 .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 유저입니다. ID: " + userId));
 
         List<Post> posts = postRepository.findByUser(user);
         return ResponseEntity.ok(posts);
     }
-
-
-
 }
